Separating flush logic in different classes and linting
